Refactor E2E chat tests by removing redundant environment setup steps and enhancing scenario descriptions. Introduce validation for Synapse environment variables in the test hooks to ensure necessary configurations are present before running tests.
